refactor: 모든 파일에 대해 매직넘버,매직스트링 상수화 함.
