+++ src/main/java/store/Constant/Discount.java
@@ -1,4 +1,4 @@
-package store.enums;
+package store.Constant;
 
 public class Discount {
     public static final double MEMBERSHIP_DISCOUNT_PERCENT = 0.3;
