+++ src/main/java/store/Constant/Error.java
@@ -1,10 +1,11 @@
-package store.enums;
+package store.Constant;
 
 public enum Error {
     ERROR_MESSAGE("[ERROR] "),
     INSUFFICIENT_STOCK_MESSAGE("[ERROR] 재고 수량을 초과하여 구매할 수 없습니다. 다시 입력해 주세요."),
     NONE_EXIST_STOCK_MESSAGE("[ERROR] 존재하지 않는 상품입니다. 다시 입력해 주세요."),
-    PURCHASE_FORMAT_MESSAGE("[ERROR] 잘못된 입력입니다. 다시 입력해 주세요.");
+    PURCHASE_FORMAT_MESSAGE("[ERROR] 잘못된 입력입니다. 다시 입력해 주세요."),
+    FILE_ERROR("파일 읽기에 실패하였습니다.");
 
     private final String errorMessage;
