+++ src/main/java/store/Constant/File.java
@@ -1,0 +1,7 @@
+package store.Constant;
+
+public class File {
+    public static final String PRODUCT_FILE_PATH = "./src/main/resources/products.md";
+    public static final String PROMOTION_FILE_PATH = "./src/main/resources/promotions.md";
+    public static final int COLUMN_HEADERS = 0;
+}
