+++ src/main/java/store/dataBase/ProductDB.java
@@ -4,17 +4,19 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
+import store.Constant.Error;
+import store.Constant.File;
 
 public class ProductDB {
 
     public List<String> findProduct() {
         while (true) {
             try {
-                List<String> input = Files.readAllLines(Paths.get("./src/main/resources/products.md"));
-                input.remove(0);
+                List<String> input = Files.readAllLines(Paths.get(File.PRODUCT_FILE_PATH));
+                input.remove(File.COLUMN_HEADERS);
                 return input;
             } catch (IOException e) {
-                System.out.println("파일 읽기에 실패하였습니다.");
+                System.out.println(Error.FILE_ERROR);
                 e.printStackTrace();
             }
         }
