+++ src/main/java/store/dataBase/PromotionDB.java
@@ -1,21 +1,22 @@
 package store.dataBase;
 
 import java.io.IOException;
-import java.nio.file.FileSystemException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.List;
+import store.Constant.Error;
+import store.Constant.File;
 
 public class PromotionDB {
 
     public List<String> findPromotion() {
         while (true) {
             try {
-                List<String> input = Files.readAllLines(Paths.get("./src/main/resources/promotions.md"));
-                input.remove(0);
+                List<String> input = Files.readAllLines(Paths.get(File.PROMOTION_FILE_PATH));
+                input.remove(File.COLUMN_HEADERS);
                 return input;
             } catch (IOException e) {
-                System.out.println("파일 읽기에 실패하였습니다.");
+                System.out.println(Error.FILE_ERROR);
                 e.printStackTrace();
             }
         }
