+++ src/main/java/store/model/StockInventory.java
@@ -48,8 +48,6 @@ public class StockInventory {
         return null;
     }
 
-    //구매할 상품종류가 재고에 존재하는지 확인하는 메소드 추가
-
     //재고가 있는지 확인하는 메소드(프로모션 + 일반재고)
     public boolean isStockAvailable(String productName, int count) {
         ProductBox normalProductBox = findNormalProductBoxByProductName(productName);
